+++ sdk/identity/Azure.Identity/tests/CredentialTestBase.cs
@@ -5,8 +5,10 @@ using System;
 using System.Globalization;
 using System.IO;
 using System.Text;
+using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
+using Azure.Core;
 using Azure.Core.TestFramework;
 using Azure.Identity.Tests.Mock;
 using Microsoft.Identity.Client;
@@ -35,10 +37,13 @@ namespace Azure.Identity.Tests
         protected AuthenticationResult result;
         protected string expectedCode;
         protected DeviceCodeResult deviceCodeResult;
-        protected const string DiscoveryResponseBody = "{\"tenant_discovery_endpoint\": \"https://login.microsoftonline.com/c54fac88-3dd3-461f-a7c4-8a368e0340b3/v2.0/.well-known/openid-configuration\",\"api-version\": \"1.1\",\"metadata\":[{\"preferred_network\": \"login.microsoftonline.com\",\"preferred_cache\": \"login.windows.net\",\"aliases\":[\"login.microsoftonline.com\",\"login.windows.net\",\"login.microsoft.com\",\"sts.windows.net\"]},{\"preferred_network\": \"login.partner.microsoftonline.cn\",\"preferred_cache\": \"login.partner.microsoftonline.cn\",\"aliases\":[\"login.partner.microsoftonline.cn\",\"login.chinacloudapi.cn\"]},{\"preferred_network\": \"login.microsoftonline.de\",\"preferred_cache\": \"login.microsoftonline.de\",\"aliases\":[\"login.microsoftonline.de\"]},{\"preferred_network\": \"login.microsoftonline.us\",\"preferred_cache\": \"login.microsoftonline.us\",\"aliases\":[\"login.microsoftonline.us\",\"login.usgovcloudapi.net\"]},{\"preferred_network\": \"login-us.microsoftonline.com\",\"preferred_cache\": \"login-us.microsoftonline.com\",\"aliases\":[\"login-us.microsoftonline.com\"]}]}";
+
+        protected const string DiscoveryResponseBody =
+            "{\"tenant_discovery_endpoint\": \"https://login.microsoftonline.com/c54fac88-3dd3-461f-a7c4-8a368e0340b3/v2.0/.well-known/openid-configuration\",\"api-version\": \"1.1\",\"metadata\":[{\"preferred_network\": \"login.microsoftonline.com\",\"preferred_cache\": \"login.windows.net\",\"aliases\":[\"login.microsoftonline.com\",\"login.windows.net\",\"login.microsoft.com\",\"sts.windows.net\"]},{\"preferred_network\": \"login.partner.microsoftonline.cn\",\"preferred_cache\": \"login.partner.microsoftonline.cn\",\"aliases\":[\"login.partner.microsoftonline.cn\",\"login.chinacloudapi.cn\"]},{\"preferred_network\": \"login.microsoftonline.de\",\"preferred_cache\": \"login.microsoftonline.de\",\"aliases\":[\"login.microsoftonline.de\"]},{\"preferred_network\": \"login.microsoftonline.us\",\"preferred_cache\": \"login.microsoftonline.us\",\"aliases\":[\"login.microsoftonline.us\",\"login.usgovcloudapi.net\"]},{\"preferred_network\": \"login-us.microsoftonline.com\",\"preferred_cache\": \"login-us.microsoftonline.com\",\"aliases\":[\"login-us.microsoftonline.com\"]}]}";
 
         public CredentialTestBase(bool isAsync) : base(isAsync)
-        { }
+        {
+        }
 
         public void TestSetup()
         {
@@ -58,7 +63,7 @@ namespace Azure.Identity.Tests
                 TenantId,
                 new MockAccount("username"),
                 null,
-                new[] { Scope },
+                new[] {Scope},
                 Guid.NewGuid(),
                 null,
                 "Bearer");
@@ -104,7 +109,7 @@ namespace Azure.Identity.Tests
                 TenantId,
                 new MockAccount("username"),
                 null,
-                new[] { Scope },
+                new[] {Scope},
                 Guid.NewGuid(),
                 null,
                 "Bearer");
@@ -151,6 +156,7 @@ namespace Azure.Identity.Tests
             {
                 return null;
             }
+
             using var memoryStream = new MemoryStream();
             request.Content.WriteTo(memoryStream, CancellationToken.None);
             memoryStream.Position = 0;
@@ -160,7 +166,8 @@ namespace Azure.Identity.Tests
             }
         }
 
-        protected MockResponse CreateMockMsalTokenResponse(int responseCode, string token, string tenantId, string userName)
+        protected MockResponse CreateMockMsalTokenResponse(int responseCode, string token, string tenantId,
+            string userName)
         {
             var response = new MockResponse(responseCode);
             var idToken = CreateMsalIdToken(Guid.NewGuid().ToString(), userName, tenantId);
@@ -191,7 +198,7 @@ namespace Azure.Identity.Tests
             return string.Format(CultureInfo.InvariantCulture, "someheader.{0}.somesignature", MsalEncode(id));
         }
 
-         private const char base64PadCharacter = '=';
+        private const char base64PadCharacter = '=';
 #if NET45
         private const string doubleBase64PadCharacter = "==";
 #endif
@@ -205,11 +212,10 @@ namespace Azure.Identity.Tests
         /// </summary>
         internal static readonly char[] s_base64Table =
         {
-            'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z',
-            'a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z',
-            '0','1','2','3','4','5','6','7','8','9',
-            base64UrlCharacter62,
-            base64UrlCharacter63
+            'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U',
+            'V', 'W', 'X', 'Y', 'Z', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p',
+            'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
+            base64UrlCharacter62, base64UrlCharacter63
         };
 
         /// <summary>
@@ -303,7 +309,7 @@ namespace Azure.Identity.Tests
                     }
                     break;
 
-                    //default or case 0: no further operations are needed.
+                //default or case 0: no further operations are needed.
             }
 
             return new string(output, 0, j);
@@ -323,6 +329,54 @@ namespace Azure.Identity.Tests
                 return null;
 
             return MsalEncode(inArray, 0, inArray.Length);
+        }
+
+        protected bool RequestBodyHasUserAssertionWithHeader(Request req, string headerName)
+        {
+            req.Content.TryComputeLength(out var len);
+            byte[] content = new byte[len];
+            var stream = new MemoryStream((int)len);
+            req.Content.WriteTo(stream, default);
+            var body = Encoding.UTF8.GetString(stream.GetBuffer(), 0, (int)stream.Length);
+            var parts = body.Split('&');
+            foreach (var part in parts)
+            {
+                if (part.StartsWith("client_assertion="))
+                {
+                    var assertion = part.AsSpan();
+                    int start = assertion.IndexOf('=') + 1;
+                    assertion = assertion.Slice(start);
+                    int end = assertion.IndexOf('.');
+                    var jwt = assertion.Slice(0, end);
+                    string convertedToken = jwt.ToString().Replace('_', '/').Replace('-', '+');
+                    switch (jwt.Length % 4)
+                    {
+                        case 2:
+                            convertedToken += "==";
+                            break;
+                        case 3:
+                            convertedToken += "=";
+                            break;
+                    }
+
+                    Utf8JsonReader reader = new Utf8JsonReader(Convert.FromBase64String(convertedToken));
+                    while (reader.Read())
+                    {
+                        if (reader.TokenType == JsonTokenType.PropertyName)
+                        {
+                            var header = reader.GetString();
+                            if (header == headerName)
+                            {
+                                return true;
+                            }
+
+                            reader.Read();
+                        }
+                    }
+                }
+            }
+
+            return false;
         }
     }
 }
